backend/requirements.txt
Django==3.2.3
Pillow==9.0.0
djangorestframework==3.12.4
djoser==2.1.0

